Remove the way we implemented probabilistic domination.
It does not work out because of missing transitivity in probablistic
domination. Consider:

    a dominates b
    b dominates c

When using a probabilistic domination function, it can be that:

    a does not dominate c

But transitivity is required for the dominance relation.

What we can do is to determine up-front (before non-dominated sorting)
for each fitness, which objective to use and which not.